refactor: Chunk 3 — split class.rs + orchestrator, deduplicate type_mapper
- Split class.rs (1048 lines) into class/ module with 5 focused files:
  mod.rs (337), constructor.rs (340), method.rs (218), routing.rs (139), mutation.rs (60)
- Split orchestrator lib.rs (508 lines) into 4 modules:
  lib.rs (58), pipeline.rs (209), scaffold.rs (190), format.rs (70)
- Deduplicate type_mapper.rs: consolidate map_ts_type/map_inner_type
  into shared map_type_core (304→257 lines)
- All files under 400 lines, all 15 tests pass, full CI pipeline green
